Move macerator components to automation package for better organization
