+++ common/src/client/java/com/logistics/automation/macerator/MaceratorRecipeBookComponent.java
@@ -1,7 +1,6 @@
-package com.logistics.core.macerator;
+package com.logistics.automation.macerator;
 
 import com.logistics.LogisticsAutomation;
-import com.logistics.LogisticsCore;
 import net.minecraft.client.gui.components.WidgetSprites;
 import net.minecraft.client.gui.screens.recipebook.GhostSlots;
 import net.minecraft.client.gui.screens.recipebook.RecipeBookComponent;
@@ -31,7 +30,7 @@ public class MaceratorRecipeBookComponent extends RecipeBookComponent<MaceratorS
     private static final Component FILTER_NAME = Component.translatable("gui.recipebook.toggleRecipes.smeltable");
 
     private static final List<TabInfo> TABS = List.of(
-        new TabInfo(LogisticsAutomation.BLOCK.MACERATOR.asItem(), LogisticsCore.RECIPE.MACERATOR_CATEGORY)
+        new TabInfo(LogisticsAutomation.BLOCK.MACERATOR.asItem(), LogisticsAutomation.RECIPE.MACERATOR_CATEGORY)
     );
 
     public MaceratorRecipeBookComponent(MaceratorScreenHandler handler) {
