+++ common/src/client/java/com/logistics/automation/macerator/jei/MaceratorRecipeCategory.java
@@ -1,8 +1,8 @@
-package com.logistics.core.macerator.jei;
+package com.logistics.automation.macerator.jei;
 
 import com.logistics.LogisticsAutomation;
 import com.logistics.LogisticsMod;
-import com.logistics.core.macerator.MaceratorRecipeWrapper;
+import com.logistics.automation.macerator.MaceratorRecipeWrapper;
 import mezz.jei.api.gui.builder.IRecipeLayoutBuilder;
 import mezz.jei.api.gui.drawable.IDrawable;
 import mezz.jei.api.gui.widgets.IRecipeExtrasBuilder;
