+++ common/src/main/java/com/logistics/LogisticsCore.java
@@ -3,22 +3,11 @@ package com.logistics;
 import com.logistics.core.LogisticsConfig;
 import com.logistics.core.bootstrap.DomainBootstrap;
 import com.logistics.core.item.WrenchItem;
-import com.logistics.core.macerator.MaceratorRecipeDisplay;
-import com.logistics.core.macerator.MaceratorRecipeSerializer;
-import com.logistics.core.macerator.MaceratorRecipeWrapper;
-import com.logistics.core.macerator.MaceratorScreenHandler;
 import com.logistics.core.lib.platform.LogisticsCreativeTab;
 import com.logistics.core.lib.platform.CreativeTabRegistrar;
 import com.logistics.core.lib.resource.ResourceId;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.crafting.RecipeBookCategory;
-import net.minecraft.world.item.crafting.RecipeSerializer;
-import net.minecraft.world.item.crafting.RecipeType;
-import net.minecraft.world.item.crafting.display.RecipeDisplay;
-import net.minecraft.world.inventory.MenuType;
-import net.minecraft.core.Registry;
-import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.world.item.CreativeModeTabs;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Items;
@@ -57,8 +46,6 @@ public final class LogisticsCore extends LogisticsMod implements DomainBootstrap
         BLOCK.register();
         ITEM.register();
         ENTITY.register();
-        MENU.register();
-        RECIPE.register();
         CREATIVE.register();
         ALIAS.register();
     }
@@ -281,53 +268,6 @@ public final class LogisticsCore extends LogisticsMod implements DomainBootstrap
 
         static void register() {
             // Macerator block entity now registered in LogisticsAutomation (moved to the automation domain).
-        }
-    }
-
-    public static final class MENU {
-        private MENU() {}
-
-        public static MenuType<MaceratorScreenHandler> MACERATOR;
-
-        static void register() {
-            MACERATOR = INSTANCE.registerMenuType("macerator", MaceratorScreenHandler::new);
-        }
-    }
-
-    public static final class RECIPE {
-        private RECIPE() {}
-
-        public static RecipeType<MaceratorRecipeWrapper> MACERATOR_RECIPE_TYPE;
-        public static RecipeSerializer<MaceratorRecipeWrapper> MACERATOR_RECIPE_SERIALIZER;
-        public static RecipeBookCategory MACERATOR_CATEGORY;
-        public static RecipeDisplay.Type<MaceratorRecipeDisplay> MACERATOR_DISPLAY_TYPE;
-
-        static void register() {
-            MACERATOR_RECIPE_TYPE = Registry.register(
-                BuiltInRegistries.RECIPE_TYPE,
-                LogisticsMod.modId("macerator").toIdentifier(),
-                new RecipeType<MaceratorRecipeWrapper>() {
-                    @Override
-                    public String toString() {
-                        return "logistics:macerator";
-                    }
-                }
-            );
-            MACERATOR_RECIPE_SERIALIZER = Registry.register(
-                BuiltInRegistries.RECIPE_SERIALIZER,
-                LogisticsMod.modId("macerator").toIdentifier(),
-                MaceratorRecipeSerializer.INSTANCE
-            );
-            MACERATOR_CATEGORY = Registry.register(
-                BuiltInRegistries.RECIPE_BOOK_CATEGORY,
-                LogisticsMod.modId("macerator").toIdentifier(),
-                new RecipeBookCategory()
-            );
-            MACERATOR_DISPLAY_TYPE = Registry.register(
-                BuiltInRegistries.RECIPE_DISPLAY,
-                LogisticsMod.modId("macerator").toIdentifier(),
-                MaceratorRecipeDisplay.TYPE
-            );
         }
     }
