+++ common/src/main/java/com/logistics/automation/macerator/MaceratorBlockEntity.java
@@ -1,4 +1,4 @@
-package com.logistics.core.macerator;
+package com.logistics.automation.macerator;
 
 import com.logistics.LogisticsAutomation;
 import com.logistics.core.machine.MachineBuilder;
