+++ fabric/src/client/java/com/logistics/LogisticsCoreClient.java
@@ -1,8 +1,6 @@
 package com.logistics;
 
 import com.logistics.core.bootstrap.ClientDomainBootstrap;
-import com.logistics.core.macerator.MaceratorScreen;
-import net.minecraft.client.gui.screens.MenuScreens;
 
 import static com.logistics.LogisticsMod.LOGGER;
 
@@ -11,8 +9,6 @@ public final class LogisticsCoreClient implements ClientDomainBootstrap {
     @Override
     public void initClient() {
         LOGGER.info("Registering core (client)");
-
-        MenuScreens.register(LogisticsCore.MENU.MACERATOR, MaceratorScreen::new);
     }
 
     @Override
