+++ fabric/src/gametest/java/com/logistics/gametest/core/MaceratorGameTest.java
@@ -3,8 +3,8 @@ package com.logistics.gametest.core;
 import com.logistics.LogisticsAutomation;
 import com.logistics.core.lib.energy.IEnergyStorage;
 import com.logistics.core.lib.storage.IItemStorage;
-import com.logistics.core.macerator.MaceratorBlock;
-import com.logistics.core.macerator.MaceratorBlockEntity;
+import com.logistics.automation.macerator.MaceratorBlock;
+import com.logistics.automation.macerator.MaceratorBlockEntity;
 import net.fabricmc.fabric.api.gametest.v1.GameTest;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
